requirements.txt
wheel
toml
numpy
jax
jaxlib
absl-py>=1.0.0
flax>=0.4.1
optax>=0.1.1
orbax-checkpoint>=0.2.1
transformers[sentencepiece]
datasets
ml_collections